+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - During global search, the search results will include all files once at most, so files within virtual directories, for example, will be excluded to prevent duplicate files.
 - The preview images when pasting an image from clipboard load faster.
 - Using the Debug menu, you can now test the new themes (Berlin, Frankfurt, and Bielefeld) during beta to help us finalise them!
+- Formatting marks at the beginning or end of a misspelled word are now excluded from the selection.
 
 ### Fixes